Created the class FlyingObject
FlyingObject is almost equal to Cannon
Deleted the line 112 "GameObject *fly = new GameObject(SCREEN_WIDTH / 6,
SCREEN_WIDTH / 6, 55, 65, 300, "media/fly.bmp", gRenderer);"
and its respectives functions on the class main.cpp

+++ cannon/Util.cpp
@@ -10,6 +10,7 @@
 
 #include <time.h>
 #include <Windows.h>
+#include <SDL.h>
 
 #else
 
@@ -17,12 +18,12 @@
 
 #endif
 
-double Util::GenerateRandom(int l, int u)
+int Util::GenerateRandom(int l, int u)
 {
     srand(time(NULL));
 
     double r = rand() % ((u - l) + 1);
     r = l + r;
 
-    return r;
+	return (int) r;
 }
